requirements.txt
streamlit
pandas
requests
xlsxwriter
Pillow